refactor: extract compare logic in a dedicated class
