+++ src/main/java/io/gravitee/policy/sslenforcement/SslEnforcementPolicy.java
@@ -26,12 +26,7 @@ import io.gravitee.policy.sslenforcement.configuration.SslEnforcementPolicyConfi
 import javax.net.ssl.SSLPeerUnverifiedException;
 import javax.net.ssl.SSLSession;
 import javax.security.auth.x500.X500Principal;
-import org.bouncycastle.asn1.ASN1ObjectIdentifier;
-import org.bouncycastle.asn1.x500.AttributeTypeAndValue;
-import org.bouncycastle.asn1.x500.RDN;
 import org.bouncycastle.asn1.x500.X500Name;
-import org.bouncycastle.asn1.x500.style.IETFUtils;
-import org.springframework.util.AntPathMatcher;
 
 /**
  * @author David BRASSELY (david.brassely at graviteesource.com)
@@ -95,7 +90,7 @@ public class SslEnforcementPolicy {
                 // Prepare name with javax.security to transform to valid bouncycastle Asn1ObjectIdentifier
                 final X500Principal x500Principal = new X500Principal(name);
                 final X500Name x500Name = new X500Name(x500Principal.getName());
-                found = areEqual(x500Name, peerName);
+                found = X500NameComparator.areEqual(x500Name, peerName);
 
                 if (found) {
                     break;
@@ -117,109 +112,5 @@ public class SslEnforcementPolicy {
         }
 
         policyChain.doNext(request, response);
-    }
-
-    private boolean areEqual(X500Name name1, X500Name name2) {
-        final RDN[] rdns1 = name1.getRDNs();
-        final RDN[] rdns2 = name2.getRDNs();
-
-        if (rdns1.length != rdns2.length) {
-            return false;
-        }
-
-        boolean reverse = false;
-
-        if (rdns1[0].getFirst() != null && rdns2[0].getFirst() != null) {
-            reverse = !rdns1[0].getFirst().getType().equals(rdns2[0].getFirst().getType()); // guess forward
-        }
-
-        for (int i = 0; i != rdns1.length; i++) {
-            if (!foundMatch(reverse, rdns1[i], rdns2)) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    private boolean foundMatch(boolean reverse, RDN rdn, RDN[] possRDNs) {
-        if (reverse) {
-            for (int i = possRDNs.length - 1; i >= 0; i--) {
-                if (possRDNs[i] != null && rDNAreEqual(rdn, possRDNs[i])) {
-                    possRDNs[i] = null;
-                    return true;
-                }
-            }
-        } else {
-            for (int i = 0; i != possRDNs.length; i++) {
-                if (possRDNs[i] != null && rDNAreEqual(rdn, possRDNs[i])) {
-                    possRDNs[i] = null;
-                    return true;
-                }
-            }
-        }
-
-        return false;
-    }
-
-    private static boolean rDNAreEqual(RDN rdn1, RDN rdn2) {
-        if (rdn1.isMultiValued()) {
-            if (rdn2.isMultiValued()) {
-                AttributeTypeAndValue[] atvs1 = rdn1.getTypesAndValues();
-                AttributeTypeAndValue[] atvs2 = rdn2.getTypesAndValues();
-
-                if (atvs1.length != atvs2.length) {
-                    return false;
-                }
-
-                for (int i = 0; i != atvs1.length; i++) {
-                    if (!atvAreEqual(atvs1[i], atvs2[i])) {
-                        return false;
-                    }
-                }
-            } else {
-                return false;
-            }
-        } else {
-            if (!rdn2.isMultiValued()) {
-                return atvAreEqual(rdn1.getFirst(), rdn2.getFirst());
-            } else {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    private static boolean atvAreEqual(AttributeTypeAndValue atv1, AttributeTypeAndValue atv2) {
-        if (atv1 == atv2) {
-            return true;
-        }
-
-        if (atv1 == null) {
-            return false;
-        }
-
-        if (atv2 == null) {
-            return false;
-        }
-
-        ASN1ObjectIdentifier o1 = atv1.getType();
-        ASN1ObjectIdentifier o2 = atv2.getType();
-
-        if (!o1.equals(o2)) {
-            return false;
-        }
-
-        String v1 = IETFUtils.canonicalize(IETFUtils.valueToString(atv1.getValue()));
-        String v2 = IETFUtils.canonicalize(IETFUtils.valueToString(atv2.getValue()));
-
-        AntPathMatcher matcher = new AntPathMatcher();
-
-        if (!matcher.match(v1, v2)) {
-            return false;
-        }
-
-        return true;
     }
 }
